Tests/LibWeb/Text/expected/aria-attribute-reflection.txt

Element Reflection for ARIA properties
Testing: role attribute reflects.
    ✅ PASS: button is equal to button as expected.
    ✅ PASS: checkbox is equal to checkbox as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-atomic attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-autocomplete attribute reflects.
    ✅ PASS: list is equal to list as expected.
    ✅ PASS: inline is equal to inline as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-braillelabel attribute reflects.
    ✅ PASS: x is equal to x as expected.
    ✅ PASS: y is equal to y as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-brailleroledescription attribute reflects.
    ✅ PASS: x is equal to x as expected.
    ✅ PASS: y is equal to y as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-busy attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-checked attribute reflects.
    ✅ PASS: mixed is equal to mixed as expected.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-colcount attribute reflects.
    ✅ PASS: 5 is equal to 5 as expected.
    ✅ PASS: 6 is equal to 6 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-colindex attribute reflects.
    ✅ PASS: 1 is equal to 1 as expected.
    ✅ PASS: 2 is equal to 2 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-colindextext attribute reflects.
    ✅ PASS: x is equal to x as expected.
    ✅ PASS: y is equal to y as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-colspan attribute reflects.
    ✅ PASS: 2 is equal to 2 as expected.
    ✅ PASS: 3 is equal to 3 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-current attribute reflects.
    ✅ PASS: page is equal to page as expected.
    ✅ PASS: step is equal to step as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-description attribute reflects.
    ✅ PASS: cold as ice is equal to cold as ice as expected.
    ✅ PASS: hot as fire is equal to hot as fire as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-disabled attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-expanded attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-haspopup attribute reflects.
    ✅ PASS: menu is equal to menu as expected.
    ✅ PASS: listbox is equal to listbox as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-hidden attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-invalid attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: grammar is equal to grammar as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-keyshortcuts attribute reflects.
    ✅ PASS: x is equal to x as expected.
    ✅ PASS: y is equal to y as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-label attribute reflects.
    ✅ PASS: x is equal to x as expected.
    ✅ PASS: y is equal to y as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-level attribute reflects.
    ✅ PASS: 1 is equal to 1 as expected.
    ✅ PASS: 2 is equal to 2 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-live attribute reflects.
    ✅ PASS: polite is equal to polite as expected.
    ✅ PASS: assertive is equal to assertive as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-modal attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-multiline attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-multiselectable attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-orientation attribute reflects.
    ✅ PASS: vertical is equal to vertical as expected.
    ✅ PASS: horizontal is equal to horizontal as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-placeholder attribute reflects.
    ✅ PASS: x is equal to x as expected.
    ✅ PASS: y is equal to y as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-posinset attribute reflects.
    ✅ PASS: 10 is equal to 10 as expected.
    ✅ PASS: 11 is equal to 11 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-pressed attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-readonly attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-relevant attribute reflects.
    ✅ PASS: text is equal to text as expected.
    ✅ PASS: removals is equal to removals as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-required attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-roledescription attribute reflects.
    ✅ PASS: x is equal to x as expected.
    ✅ PASS: y is equal to y as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-rowcount attribute reflects.
    ✅ PASS: 10 is equal to 10 as expected.
    ✅ PASS: 11 is equal to 11 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-rowindex attribute reflects.
    ✅ PASS: 1 is equal to 1 as expected.
    ✅ PASS: 2 is equal to 2 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-rowindextext attribute reflects.
    ✅ PASS: x is equal to x as expected.
    ✅ PASS: y is equal to y as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-rowspan attribute reflects.
    ✅ PASS: 2 is equal to 2 as expected.
    ✅ PASS: 3 is equal to 3 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-selected attribute reflects.
    ✅ PASS: true is equal to true as expected.
    ✅ PASS: false is equal to false as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-setsize attribute reflects.
    ✅ PASS: 10 is equal to 10 as expected.
    ✅ PASS: 11 is equal to 11 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-sort attribute reflects.
    ✅ PASS: descending is equal to descending as expected.
    ✅ PASS: ascending is equal to ascending as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-valuemax attribute reflects.
    ✅ PASS: 99 is equal to 99 as expected.
    ✅ PASS: 100 is equal to 100 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-valuemin attribute reflects.
    ✅ PASS: 3 is equal to 3 as expected.
    ✅ PASS: 2 is equal to 2 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-valuenow attribute reflects.
    ✅ PASS: 50 is equal to 50 as expected.
    ✅ PASS: 51 is equal to 51 as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
Testing: aria-valuetext attribute reflects.
    ✅ PASS: 50% is equal to 50% as expected.
    ✅ PASS: 51% is equal to 51% as expected.
    ✅ PASS: null is equal to null as expected.
    ✅ PASS: null is equal to null as expected.
